requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenAI
openai==1.3.5

# Vector Store
faiss-cpu==1.7.4
numpy==1.24.3

# Document Processing
pymupdf==1.23.8
reportlab==4.0.7

# Web Automation
selenium==4.15.2
webdriver-manager==4.0.1

# Data Processing
pandas==2.1.3
requests==2.31.0

# GUI
gradio==4.7.1

# Utilities
python-dotenv==1.0.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
